third_party/blink/web_tests/http/tests/security/powerfulFeatureRestrictions/old-powerful-features-on-insecure-origin-expected.txt

CONSOLE WARNING: line 20: getCurrentPosition() and watchPosition() no longer work on insecure origins. To use this feature, you should consider switching your application to a secure origin, such as HTTPS. See https://goo.gl/rStTGz for more details.
This is a testharness.js-based test.
PASS getCurrentPosition
PASS watchPosition
PASS Media devices properties
PASS requestMediaKeySystemAccess
PASS applicationCache
Harness: the test ran to completion.
